Fix some wonky stuff
